agent/requirements.txt
openai
python-dotenv
jupyter-client 